chore: clean up codebase — remove dead code, minimize comments, deduplicate utilities
- Remove 96-line README-style header from classification-provider.js, replace with 3-line summary
- Remove ~30 redundant JSDoc comments and inline comments across all packages
- Remove deprecated batchUpsert() method and MemoryBatchUpsertRequest type from cortex-client
- Remove duplicate plugin-sdk.d.ts (identical copy of plugin-sdk.ts) from openclaw-memory
- Extract shared cortexFetch helper from 3 duplicate implementations in cortex-mcp-server tools
- Convert synapse.js static Anthropic import to dynamic import (matches memory.js pattern)
- Fix hardcoded version strings (0.1.0, 0.2.0) to match package.json (1.0.1)
- Update openclaw-memory to use bulkStore() instead of removed batchUpsert()
- Update test from batchUpsert to bulkStore

All 328 unit tests pass. Integration-tested all endpoints against live Cortex instance:
Cortex core (12 endpoints), cortex-client (12 methods), MCP server (8 tools), openclaw-memory (19 tests).
